Feature/add pumping out

+++ src/main/java/com/diplom/padding/repositories/app/RoleRepository.java
@@ -3,5 +3,5 @@ package com.diplom.padding.repositories.app;
 import com.diplom.padding.entity.app.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Byte> {
 }
